pyproject.toml
[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[project]
name = "urban-mapper"
description = "🗺️ Spatial Join & Enrich any urban layer given any external urban dataset of interest, streamline your urban analysis with Scikit-Learn-Like pipelines, and share your insights with the urban research community!"
authors = [
     {name = "Provost Simon", email = "sgp29@kent.ac.uk"}, # First author & maintainer
     {name = "New York University", email = "fairuse@nyu.edu"}, # Copyright holder
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.10"
keywords = []
version = "0.1.5"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Topic :: Documentation",
    "Topic :: Software Development",
    "Topic :: Utilities",
    "Typing :: Typed",
]
dependencies = [
    "folium>=0.19.4",
    "ipython>=8.18.1",
    "ipywidgets>=8.1.5",
    "matplotlib>=3.9.4",
    "networkx>=3.2.1",
    "scikit-learn>=1.6.1",
    "beartype>=0.19.0",
    "osmnx>=2.0.1",
    "geopandas>=1.0.1",
    "dependency-injector>=4.45.0",
    "pyarrow>=19.0.1",
    "loguru>=0.7.3",
    "thefuzz>=0.22.1",
    "dill<0.3.8",
    "geopy>=2.4.1",
    "alive-progress>=3.2.0",
    "datasets>=3.5.0",
    "mapclassify>=2.8.1",
    "h3>=4.3.1",
]

[project.optional-dependencies]
notebook = [
    "jupyter>=1.1.1",
    "notebook==6.4.12",
    "jupyter-server-ydoc>=1.1.0",
    "jupyter-contrib-nbextensions>=0.7.0",
    "jupyterlab-widgets>=3.0.13",
    "jupyter-resource-usage>=1.1.1",
    "nbstripout>=0.8.1",
]
interactive_table_vis = [
    "skrub>=0.5.1",
]
auctus_mixins = [
    "auctus-search",
]
pipeline_generators = [
    "ell-ai[all]>=0.0.17",
]
jupytergis_mixins = [
    "jupytergis==0.4.4",
    "jupytergis-core>=0.2.1",
    "jupytergis-qgis>=0.2.1",
]
full = [
    "skrub>=0.5.1",
    "auctus-search",
    "ell-ai[all]>=0.0.17",
    "jupyter>=1.1.1",
    "notebook==6.4.12",
    "jupyter-server-ydoc>=1.1.0",
    "jupyter-contrib-nbextensions>=0.7.0",
    "jupyterlab-widgets>=3.0.13",
    "jupyter-resource-usage>=1.1.1",
    "nbstripout>=0.8.1",
    "jupytergis==0.4.4",
    "jupytergis-core>=0.2.1",
    "jupytergis-qgis>=0.2.1",
]

[project.urls]
Homepage = "https://github.com/VIDA-NYU/UrbanMapper"
Documentation = "https://urbanmapper.readthedocs.io/en/latest/"
Changelog = "https://github.com/VIDA-NYU/UrbanMapper/blob/main/CHANGELOG.md"
Repository = "https://github.com/VIDA-NYU/UrbanMapper"
Issues = "https://github.com/VIDA-NYU/UrbanMapper/issues"
Discussions = "https://github.com/VIDA-NYU/UrbanMapper/discussions"

[tool.pdm.build]
# Include as much as possible in the source distribution, to help redistributors.
excludes = ["**/.pytest_cache"]
source-includes = [
    "config",
    "*.md",
    "LICENSE",
]

[dependency-groups]
dev = [
    # maintenance
    "build>=1.2",
    "git-changelog>=2.5",
    "twine>=5.1",
    # ci
    "duty>=1.4",
    "ruff>=0.4",
    "pytest>=8.2",
    "pytest-cov>=5.0",
    "pytest-randomly>=3.15",
    "pytest-xdist>=3.6",
    "mypy>=1.10",
    "types-markdown>=3.6",
    "types-pyyaml>=6.0",
    # docs
    "black>=24.4",
    "markdown-callouts>=0.4",
    "markdown-exec>=1.8",
    "mkdocs>=1.6",
    "mkdocs-coverage>=1.0",
    "mkdocs-gen-files>=0.5",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "mkdocs-literate-nav>=0.6",
    "mkdocs-minify-plugin>=0.8",
    "mkdocstrings[python]>=0.25",
    # YORE: EOL 3.10: Remove line.
    "tomli>=2.0; python_version < '3.11'",
    "mkdocstrings-python>=1.16.10",
    "mkdocs-jupyter>=0.25.1",
    "pymdown-extensions>=10.14.3",
    "mkdocs-material[emoji,imaging]>=9.6.11",
    "mkdocs-mermaid2-plugin>=1.2.1",
    "mkdocs-macros-plugin>=1.3.7",
    "mkdocs-abs-rel-plugin>=0.2.4",
    "mkdocs-document-dates>=3.1.5",
    # development workflow helpers
    "pre-commit>=4.1.0",
]

notebook = [
    "jupyter>=1.1.1",
    "notebook==6.4.12",
    "jupyter-server-ydoc>=1.1.0",
    "jupyter-contrib-nbextensions>=0.7.0",
    "jupyterlab-widgets>=3.0.13",
    "jupyter-resource-usage>=1.1.1",
    "nbstripout>=0.8.1",
]

interactive_table_vis = [
    "skrub>=0.5.1",
]

auctus_mixins = [
    "auctus-search",
]

pipeline_generators = [
    "ell-ai[all]>=0.0.17",
]

jupytergis_mixins = [
    "jupytergis==0.4.4",
    "jupytergis-core>=0.2.1",
    "jupytergis-qgis>=0.2.1",
]

full = [
    "skrub>=0.5.1",
    "auctus-search",
    "ell-ai[all]>=0.0.17",
    "jupyter>=1.1.1",
    "notebook==6.4.12",
    "jupyter-server-ydoc>=1.1.0",
    "jupyter-contrib-nbextensions>=0.7.0",
    "jupyterlab-widgets>=3.0.13",
    "jupyter-resource-usage>=1.1.1",
    "nbstripout>=0.8.1",
    "jupytergis==0.4.4",
    "jupytergis-core>=0.2.1",
    "jupytergis-qgis>=0.2.1",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
